refac disabled tests
